requirements.txt
numpy
setuptools
PyYAML
fastapi
uvicorn
./rpi-ws2805-python/library
